+++ ios/Classes/UIImage/TiUIImage+Rotate.h
@@ -1,0 +1,15 @@
+//
+//  UIImage+Rotate.h
+//  imagefactory
+//
+//  Created by Collin Price on 2015-07-06.
+//
+//
+
+#import <UIKit/UIKit.h>
+
+@interface TiUIImageRotate : NSObject
+
++ (UIImage *)rotateImage:(CGFloat)degrees image:(UIImage*)image;
+
+@end