+++ test/unit/index.js
@@ -1,0 +1,3 @@
+const testsContext = require.context('./specs', true, /\.spec\.js$/);
+
+testsContext.keys().forEach(testsContext);